добавил функции report
